+++ src/Exceptions/NoSuchRoleException.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace TPG\Deadbolt\Exceptions;
+
+use Throwable;
+
+class NoSuchRoleException extends \Exception
+{
+    public function __construct(string $role)
+    {
+        $message = 'The role '.$role.' does not exist';
+        parent::__construct($message, 1);
+    }
+}