+++ src/Traits/Deadbolted.php
@@ -2,14 +2,14 @@
 
 namespace TPG\Deadbolt\Traits;
 
-use TPG\Deadbolt\Permissions;
+use TPG\Deadbolt\Deadbolt;
 
 trait Deadbolted
 {
     protected $deadbolt;
 
-    public function deadbolt(): Permissions
+    public function deadbolt(): Deadbolt
     {
-        return $this->deadbolt ?: $this->deadbolt = new Permissions($this, config('deadbolt'));
+        return $this->deadbolt ?: $this->deadbolt = new Deadbolt($this, config('deadbolt'));
     }
 }